relacionamos la orden de compra con la compra, y agregamos la vigencia del timbrado de la compra y de la nota de crédito de la compra

+++ database/migrations/2018_12_12_001354_alter_notas_credito_compras_cab_table.php
@@ -15,6 +15,7 @@ class AlterNotasCreditoComprasCabTable extends Migration
     {
         Schema::table('nota_credito_compras_cab', function (Blueprint $table) {
             $table->integer('compra_cab_id')->unsigned();
+            $table->date('fecha_vigencia_timbrado');
 
             $table->foreign('compra_cab_id')->references('id')->on('compras_cab');
         });
@@ -27,8 +28,18 @@ class AlterNotasCreditoComprasCabTable extends Migration
      */
     public function down()
     {
-        Schema::table('nota_credito_compras_cab', function (Blueprint $table) {
-            $table->dropColumn('compra_cab_id');
-        });
+        //Pablo - 14/10/2018 - controlamos si existen las columnas primero antes del drop        
+        if(Schema::hasColumn('nota_credito_compras_cab','compra_cab_id') ) {
+            Schema::table('nota_credito_compras_cab', function (Blueprint $table) {
+                $table->dropColumn('compra_cab_id');
+            });
+        }
+
+        //Pablo - 14/10/2018 - controlamos si existen las columnas primero antes del drop        
+        if(Schema::hasColumn('nota_credito_compras_cab','fecha_vigencia_timbrado') ) {
+            Schema::table('nota_credito_compras_cab', function (Blueprint $table) {
+                $table->dropColumn('fecha_vigencia_timbrado');
+            });
+        }
     }
 }
